refactor: ProductRepository의 추상화
