+++ README.md
@@ -7,4 +7,4 @@
 Repository 객체를 교체하기 위해 또다른 객체를 만들어 바꿔줘야 한다. 즉, ViewModel에 직접적인 변경사항이 발생한다.
 
 - [x] CartRepository의 추상화
-- [ ] ProductRepository의 추상화
+- [x] ProductRepository의 추상화
