+++ app/src/main/java/woowacourse/shopping/ui/MainActivity.kt
@@ -8,7 +8,7 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
 import woowacourse.shopping.data.DefaultCartRepository
-import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.data.DefaultProductRepository
 import woowacourse.shopping.databinding.ActivityMainBinding
 import woowacourse.shopping.ui.cart.CartActivity
 
@@ -18,7 +18,7 @@ class MainActivity : AppCompatActivity() {
 
     private val viewModel by viewModels<MainViewModel> {
         MainViewModelFactory(
-            ProductRepository(),
+            DefaultProductRepository(),
             DefaultCartRepository()
         )
     }
